requirements.txt
apex==0.9.10dev
einops==0.7.0
gensim==4.3.2
h5py==3.8.0
imageio==2.31.3
info_nce_pytorch==0.1.4
lmdb==1.4.1
matplotlib==3.7.1
numpy==1.23.5
opencv_python==4.7.0.72
opencv_python_headless==4.8.1.78
pandas==2.0.1
peft==0.8.2
Pillow==9.3.0
Pillow==10.3.0
scikit_learn==1.2.2
scipy==1.13.0
skimage==0.0
timm==0.9.12
torch==2.0.0+cu118
torchvision==0.15.1+cu118
tqdm==4.65.0
transformers==4.35.0.dev0
wfdb==3.2.0
